+++ Wire-iOS/Sources/Helpers/URL+String.swift
@@ -16,8 +16,30 @@
 //
 
 import Foundation
+import WireSystem
+
+private let zmLog = ZMSLog(tag: "URL Helper")
 
 extension URL {
+    func decode<T>(_ type: T.Type) throws -> T where T : Decodable {
+        let data: Data
+        do {
+            data = try Data(contentsOf: self)
+        } catch {
+            zmLog.error("Failed to load \(type) at path: \(self), error: \(error)")
+            throw error
+        }
+        
+        let decoder = JSONDecoder()
+        
+        do {
+            return try decoder.decode(type, from: data)
+        } catch {
+            zmLog.error("Failed to parse JSON at path: \(self), error: \(error)")
+            throw error
+        }
+    }
+
     var urlWithoutScheme: String {
         return stringWithoutPrefix("\(scheme ?? "")://")
     }
